% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_shapes.R
\name{add_fishape}
\alias{add_fishape}
\title{fish silhouette in ggplot2}
\usage{
add_fishape(
  family = "Acanthuridae",
  option = NA,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  fill = "#000000"
)
}
\arguments{
\item{family}{A character string indicating the fish family to use.}

\item{option}{A character string indicating the fish species to use.
If NA the first avalable option within a family will be used}

\item{xmin}{x location (in data coordinates) giving horizontal location of raster}

\item{xmax}{x location (in data coordinates) giving horizontal location of raster}

\item{ymin}{y location (in data coordinates) giving vertical location of raster}

\item{ymax}{y location (in data coordinates) giving vertical location of raster}
}
\description{
Adds a fish silhouette to your plot
}
\examples{
library(ggplot2)

ggplot() + add_fishape(fill = "#2f87e4")

ggplot(diamonds)+
  geom_bar(aes(cut, fill = cut)) +
  scale_fill_fish_d(option = "Naso_lituratus") +
  add_fishape(family = "Acanthuridae",
              option = "Naso_unicornis",
              xmin = 1, xmax = 3, ymin = 15000, ymax = 20000,
              fill = fish(option = "Naso_lituratus", n = 4)[2]) +
  theme_bw()

}
