% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pallette.R
\name{add_silhouette}
\alias{add_silhouette}
\title{fish silhouette in ggplot2}
\usage{
add_silhouette(option = "Balistapus_undulatus", xmin = -Inf,
  xmax = Inf, ymin = -Inf, ymax = Inf)
}
\arguments{
\item{option}{A character string indicating the fish species to use.}

\item{xmin}{x location (in data coordinates) giving horizontal location of raster}

\item{xmax}{x location (in data coordinates) giving horizontal location of raster}

\item{ymin}{y location (in data coordinates) giving vertical location of raster}

\item{ymax}{y location (in data coordinates) giving vertical location of raster}
}
\description{
Adds a fish silhouette to your plot
}
\examples{
library(ggplot2)

ggplot() + add_silhouette()

ggplot(diamonds)+
  geom_bar(aes(cut, fill = cut)) +
  scale_fill_fish_d(option = "Balistapus_undulatus") +
  add_silhouette(option = "Balistapus_undulatus", ymin = 15000, ymax = 20000, xmin = 2, xmax = 3) +
  theme_bw()

}
