% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pallette.R
\name{fish_palettes}
\alias{fish_palettes}
\title{Available Palettes.}
\usage{
fish_palettes()
}
\value{
\code{fish_palettes} returns a character vector with the names of the fish palettes available to use.
}
\description{
This function returns a vector containing the names of all the available palettes in the 'fishualize' package.
}
\examples{
fish_palettes()

}
