% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_shapes.R
\name{fishapes}
\alias{fishapes}
\title{Available fish silhouettes}
\usage{
fishapes()
}
\value{
\code{fishapes} returns a dataframe containing the all the available fish
silhouettes available to use.
}
\description{
This function returns a dataframe containing the all the available fish
silhouettes accessible through the 'fishualize' package.
}
\examples{
fishapes()

}
