% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pallette.R
\docType{data}
\name{fishcolors}
\alias{fishcolors}
\title{Original fish color database}
\format{A data frame containing all the colours used in the palette:
\itemize{
  \item option: It is intended to be a general option for choosing the specific colour palette.
  \item hex: hex color code
}}
\usage{
fishcolors
}
\description{
A dataset containing some colour palettes inspired by fish species
}
\keyword{datasets}
