% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pallette.R
\name{fishualize}
\alias{fishualize}
\title{Visualization of fish color palette}
\usage{
fishualize(option = "Centropyge_loricula", n = 5, ...)
}
\arguments{
\item{option}{A character string indicating the fish species to use.}

\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{...}{Other arguments as can be specified in the function \code{fish}.
See ?fishualize::fish for details.}
}
\value{
\code{fishualize} returns a visualisation of the specified color palette.
}
\description{
This function creates an image of the specified fish color palette.
}
\examples{

fishualize::fishualize()
fishualize::fishualize(option = "Zanclus_cornutus", n = 8)

}
